@@ -1,21 +1,14 @@
-import Vue from 'vue'
-import VueI18n from 'vue-i18n'
-
+import { boot } from 'quasar/wrappers'
+import { createI18n } from 'vue-i18n'
 import messages from 'src/i18n'
 
-Vue.use(VueI18n)
+export default boot(({ app }) => {
+  const i18n = createI18n({
+    locale: 'en-US',
+    globalInjection: true,
+    messages,
+  })
 
-const i18n = new VueI18n({
-  locale: 'en-us',
-  fallbackLocale: 'en-us',
-  messages
-})
-
-export default ({ app }) => {
   // Set i18n instance on app
-  app.i18n = i18n
-}
-
-// if you need to import it from
-// other files, then:
-export { i18n }
+  app.use(i18n)
+})
