@@ -2,41 +2,34 @@
   <section id="work" class="container">
     <h2>{{ $t('work.word1') }} <span>{{ $t('work.word2') }}</span></h2>
     <q-timeline :layout="timelineLayout">
-      <q-timeline-entry
-        :title="$t('work.timeline1.title') + ''"
-        :subtitle="$t('work.timeline1.subtitle')"
-        side="right"
-      >
+      <q-timeline-entry :title="$t('work.timeline1.title') + ''" :subtitle="$t('work.timeline1.subtitle')" side="right">
         <h6 class="location">{{ $t('work.timeline1.location') }}</h6>
         <div>
           {{ $t('work.timeline1.div') }}
           <ul>
-            <li><a @click="projectLinkClick('KraptTaskManager')" class="dialog__link">KraptTaskManager</a> - {{ $t('work.timeline1.li1') }}</li>
+            <li><a @click="updateProjectDialog('KraptTaskManager')" class="dialog__link">KraptTaskManager</a> -
+              {{ $t('work.timeline1.li1') }}</li>
             <li>{{ $t('work.timeline1.li2') }}</li>
-            <li><a @click="projectLinkClick('PublishNews')" class="dialog__link">PublishNews</a> - {{ $t('work.timeline1.li3') }}</li>
-            <li><a @click="projectLinkClick('Dobrodey11')" class="dialog__link">Dobrodey11</a> - {{ $t('work.timeline1.li4') }} <a href="http://krapt-rk.ru/apk/dobrodey11.apk">http://krapt-rk.ru/apk/dobrodey11.apk</a></li>
-            <li><a @click="projectLinkClick('AdminPanelDobrodey11')" class="dialog__link">AdminPanelDobrodey11</a> - {{ $t('work.timeline1.li5') }}</li>
+            <li><a @click="updateProjectDialog('PublishNews')" class="dialog__link">PublishNews</a> -
+              {{ $t('work.timeline1.li3') }}</li>
+            <li><a @click="updateProjectDialog('Dobrodey11')" class="dialog__link">Dobrodey11</a> -
+              {{ $t('work.timeline1.li4') }} <a
+                href="http://krapt-rk.ru/apk/dobrodey11.apk">http://krapt-rk.ru/apk/dobrodey11.apk</a></li>
+            <li><a @click="updateProjectDialog('AdminPanelDobrodey11')" class="dialog__link">AdminPanelDobrodey11</a> -
+              {{ $t('work.timeline1.li5') }}</li>
             <li>{{ $t('work.timeline1.li6') }}</li>
           </ul>
         </div>
       </q-timeline-entry>
 
-      <q-timeline-entry
-        :title="$t('work.timeline2.title')"
-        :subtitle="$t('work.timeline2.subtitle')"
-        side="left"
-      >
+      <q-timeline-entry :title="$t('work.timeline2.title')" :subtitle="$t('work.timeline2.subtitle')" side="left">
         <h6 class="location">{{ $t('work.timeline2.location') }}</h6>
         <div>
           {{ $t('work.timeline2.div') }}
         </div>
       </q-timeline-entry>
 
-      <q-timeline-entry
-        :title="$t('work.timeline3.title')"
-        :subtitle="$t('work.timeline3.subtitle')"
-        side="right"
-      >
+      <q-timeline-entry :title="$t('work.timeline3.title')" :subtitle="$t('work.timeline3.subtitle')" side="right">
         <h6 class="location">{{ $t('work.timeline3.location') }}</h6>
         <div>
           {{ $t('work.timeline3.div') }}
@@ -59,49 +52,38 @@
   </section>
 </template>
 
-<script>
+<script setup>
+import { computed, ref } from 'vue'
+import { useQuasar } from 'quasar'
 import { fasTimes } from '@quasar/extras/fontawesome-v5'
 
-export default {
-  name: 'Work',
+const dialog = ref(false)
+const imageDialog = ref('')
 
-  created () {
-    this.fasTimes = fasTimes
-  },
+const imagesDialog = {
+  KraptTaskManager: 'images/KraptTaskManager.png',
+  PublishNews: 'images/PublishNews.png',
+  Dobrodey11: 'images/Dobrodey11.png',
+  AdminPanelDobrodey11: 'images/AdminPanelDobrodey11.png'
+}
 
-  data: () => ({
-    dialog: false,
-    imageDialog: '',
-    imagesDialog: {
-      KraptTaskManager: 'images/KraptTaskManager.png',
-      PublishNews: 'images/PublishNews.png',
-      Dobrodey11: 'images/Dobrodey11.png',
-      AdminPanelDobrodey11: 'images/AdminPanelDobrodey11.png'
-    }
-  }),
+const $q = useQuasar()
 
-  computed: {
-    timelineLayout () {
-      if (this.$q.platform.is.mobile) {
-        return 'dense'
-      } else {
-        return 'loose'
-      }
-    },
-    widthDialog () {
-      return window.innerWidth >= 460 ? 400 : window.innerWidth - 60
-    },
-    titleDialog () {
-      return this.imageDialog.length > 0 ? this.imageDialog.split('/')[1].split('.')[0] : ''
-    }
-  },
+const timelineLayout = computed(() => {
+  return $q.platform.is.mobile ? 'dense' : 'loose'
+})
 
-  methods: {
-    projectLinkClick (name) {
-      this.imageDialog = this.imagesDialog[name]
-      this.dialog = true
-    }
-  }
+const widthDialog = computed(() => {
+  return window.innerWidth >= 460 ? 400 : window.innerWidth - 60
+})
+
+const titleDialog = computed(() => {
+  return imageDialog.value.length > 0 ? imageDialog.value.split('/')[1].split('.')[0] : ''
+})
+
+function updateProjectDialog(name) {
+  imageDialog.value = imagesDialog[name]
+  dialog.value = true
 }
 </script>
 
@@ -120,7 +102,10 @@ export default {
     list-style-type: disc;
   }
 
-  a:link, a:visited, a:focus, a:active {
+  a:link,
+  a:visited,
+  a:focus,
+  a:active {
     color: $primary;
   }
 
