@@ -1,7 +1,7 @@
-import enUS from './en-us'
+import enUS from './en-US'
 import ru from './ru'
 
 export default {
-  'en-us': enUS,
-  ru: ru
+  'en-US': enUS,
+  ru,
 }
